lab10.2

+++ lab11_1.cpp
@@ -1,7 +1,7 @@
 #include<iostream>
 using namespace std;
 int x;
-int fibonacci(int x);
+long long int fibonacci(int x);
 
 
 int main(){
@@ -11,7 +11,7 @@ int main(){
     return 0; 
 }
 
-int fibonacci (int x){
+long long int fibonacci (int x){
     if(x==0)
     return 0;
     if (x==1)
